Started scraping data
Changed the game_results.csv to tab-delimited for ease of browse.
Scraped the player statistics for Konstantin Okulov to be used as a sample for adjusting the script. Problems to be solved: table headers not recognised as such, player name not included in any of the stats tables.
Saved Konstantin Okulov's statistics as separate .csv files to work on locally.