Gathered player profile links
Finished the script for scraping a list of player profile links from a webpage into a single dataframe. However, the used webpage turned out to contain a hidden filter showing only the players whose surnames start with the "A" letter. Next step - iterate through all letters of English alphabet to gather a complete list of profile links for all KHL players in history.