Removed data from Kaggle
The focus of the project has completely shifted towards scrapping data from the KHL website directly. As such, there is no need to use the Kaggle dataset anymore, particularly because it has not been kept up-to-date. Adding the results of the new matches would be almost as much work as gathering all data from the scratch.
In addition, such an approach allows us to avoid complications when trying to integrate the two different datasets.