Made a final notebook for Stage 1
FIRST BIG KEYSTONE

Combined players_scraping_list.ipynb, players_scraping_data.ipynb and players_cleaning_data.ipynb into a single notebook that gets the data on all players from the KHL website. The current version is quite messy and so will be better structured later on.

Since this is a test run, the notebook was modified to send out a rather limited number of web requests. Specifically, it scrapes the profile links for players whose surnames start with either "A" or "B" letter and the data for the first 5 players from each, for a total of 12 requests.