Finished Stage 1 notebook
Made the notebook into its final format. The code is now broken up into separate functions and presented in a much cleaner and easy-to-understand way (I hope).

The data for all players is not yet uploaded as the whole notebook has not been run yet except for the test runs on a small sample set of players. It will be added in a next commit or two.

There might not be any updates to the project for a day or two as I am currently figuring out how to use AWS (Amazon Web Service) for web scraping. It might be totally unnecessary at this point since the number of web scraping requests is fairly low (3364 players to be exact) but it would be a useful instrument to have on hand for the future.