Renamed Stage 1 files
Stage 2 will deal with data processing, and we want to be able to clearly tell apart the original raw data from the processed one. To this end, prefix "raw_" was added to all data set for processing.

Streamlined the file names to the same standard of length. This could help organising the project's files in the future as it grows in size.