Finished processing career statistics
The data for skaters and goalies is now saved in two separate files. The tables only have columns relevant to that role and store most of the values as integers rather than floats.

In addition, icetime was recalculated in seconds and stored as integers next to the original column stored in the "minutes:seconds" format.