Working on match statistics
Did everything except for fixing the goalie icetime. Some goalies already have icetime in seconds instead of the 'minutes:seconds' format, and while I have a function that is supposed to fix it nothing seems to work. Particularly, Pandas apply function returns a Series of tuples instead of a tuple of Series and that breaks everything.