Updated README
Added the description of everything done over the course of Stage 2 - data processing. It can be found in the "Completed" part of the project life cycle.

Described the idea of what could be an interesting Machine Learning analysis on team-level match data. The approach would be focused around dummy variables for players. It can be found in the "Further plans" part of the project life cycle.