Completed changes to Stage 2 notebooks
The streamlining and styling changes from match data processing notebook have been applied to the rest of Stage 2 notebooks.

Renamed the column identifying the player as either a skater or a goalie from "Role" to "Position". The "Role" column is now a column identifying the player as either a captain, assistant captain or usual player.

Fixed the warnings caused by using views instead of the copies of the dataframes.