Minor fixes
A match between Slovan and Vityaz teams on January 11th 2019 had proper statistics except icetime for all skaters. Therefore, after dropping the rows without proper icetime only goalies were left in that match, and the number of match IDs for skaters and goalies did not match. For the sake of simplicity, that match was dropped from the observations.