Started analysing data
The first approach will be creating a match-level data with dummies for season, team and every player that has participated in that match. The dummy variables will be equal to +1 if it is a home team or a home team player and -1 if it is an away team or an away team player, 0 for the uninvolved dummies. That way, we can teach the model to take into account both opposing teams and their rosters.