Added test notebooks
Most of them are probably going to be deleted later down the line, but I want to store them for now. The notebooks are mainly slight alternations of the original Dummies.ipynb notebook, although the "Test" family of notebooks is generally using a somewhat different approach than the "Dummies" one.